Awaylower plugin - set nicks to derank when used
Entirely optional plugin; Add an awaynick and a user will be stripped of
every channel rank that isn't voice in channels the bot is op in. When
they nick out of it, they get their ranks back, provided the bot is
still op.

For security reasons, the ranks are not saved, and are also cleared when
a user parts a channel or quits. Voice is not removed, nor given back.